refactor!: the package and the entity ID are now elmo_iess_alarm
