+++ pyproject.toml
@@ -6,9 +6,9 @@ requires = [
 build-backend = "hatchling.build"
 
 [project]
-name = "econnect-alarm"
+name = "elmo-iess-alarm"
 dynamic = ["version"]
-description = "Home Assistant integration that provides a full-fledged Alarm Panel to control your Elmo alarm system."
+description = "Home Assistant integration that provides a full-fledged Alarm Panel to control your Elmo/IESS alarm systems."
 readme = "README.md"
 requires-python = ">=3.11"
 license = "Apache-2.0"
@@ -19,6 +19,8 @@ keywords = [
   "alarm-system",
   "econnect",
   "elmo",
+  "metronet",
+  "iess",
 ]
 authors = [
   { name = "Emanuele Palazzetti", email = "emanuele.palazzetti@gmail.com" },
@@ -58,8 +60,8 @@ lint = [
 ]
 
 all = [
-  "econnect-alarm[dev]",
-  "econnect-alarm[lint]",
+  "elmo-iess-alarm[dev]",
+  "elmo-iess-alarm[lint]",
 ]
 
 [project.urls]
@@ -68,7 +70,7 @@ Issues = "https://github.com/palazzem/ha-econnect-alarm/issues"
 Source = "https://github.com/palazzem/ha-econnect-alarm"
 
 [tool.hatch.version]
-path = "custom_components/econnect_alarm/manifest.json"
+path = "custom_components/elmo_iess_alarm/manifest.json"
 pattern = '"version":\s"(?P<version>[^"]+)"'
 
 [tool.hatch.metadata]
@@ -78,7 +80,7 @@ allow-direct-references = true
 asyncio_mode = "auto"
 
 [tool.hatch.build.targets.sdist]
-only-include = ["custom_components/econnect_alarm"]
+only-include = ["custom_components/elmo_iess_alarm"]
 
 [tool.hatch.build.targets.zipped-directory]
-only-include = ["custom_components/econnect_alarm"]
+only-include = ["custom_components/elmo_iess_alarm"]
